feat: refactored handlers for general forms
