+++ src/components/molecule/AuthFormMolecule.jsx
@@ -1,0 +1,9 @@
+const AuthFormMolecule = () => {
+  return (
+    <div>
+      <h1>Auth</h1>
+    </div>
+  );
+};
+
+export default AuthFormMolecule;
