+++ src/components/organism/FormOrganism.jsx
@@ -1,309 +1,14 @@
-import Form from "react-bootstrap/Form";
-import { useState } from "react";
-
-import styles from "../../styles/organism/Form.module.css";
-import { buttonClick, subheading } from "../atom/elements";
-import AxiosInstance from "../../API/AxiosInstance";
-import useAppContext from "../../hooks/useAppContext";
+// components
+import AuthFormMolecule from "../molecule/AuthFormMolecule";
+import GeneralFormMolecule from "../molecule/GeneralFormMolecule";
 
 /**
  * Displaying of form data based on state store
  * @returns {JSX.Element}
  * @constructor
  */
-const FormOrganism = () => {
-  const { state, dispatch } = useAppContext();
-  const { formsReducers } = state;
-  const { form } = formsReducers;
-
-  console.log("form", form);
-
-  const [formData, setFormData] = useState({});
-
-  const handleChange = (e) => {
-    e.preventDefault();
-
-    const { name, value } = e.target;
-
-    setFormData({
-      ...formData,
-      [name]: value,
-    });
-  };
-
-  // const handleImage = (e) => {
-  //   e.preventDefault();
-  //
-  //   setFormData({
-  //     ...formData,
-  //     property_image: e.target.files,
-  //   });
-  // };
-
-  // const handleImage = (e) => {
-  //   let form = new FormData();
-  //
-  //   if (e.image_url) {
-  //     form.append("image_url", e.image_url);
-  //     e.image_url.name;
-  //   }
-  // };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    dispatch({ type: "TOGGLE HIDE MODAL" });
-
-    await AxiosInstance.post(form.SubmitURL, formData, {
-      headers: {
-        "content-type": "multipart/form-data",
-      },
-    })
-      .then((res) => {
-        console.log("res", res);
-        dispatch({
-          type: form.Success.update_user,
-          payload: form.Success.payload ? res.data : null,
-        });
-        dispatch({ type: "SHOW ALERT MESSAGE", payload: form.Success.message });
-      })
-      .catch((err) => {
-        console.log("err", err);
-        dispatch({ type: "ERROR UPDATING USER DATA", payload: err.message });
-      });
-  };
-
-  const getBase64 = (file) => {
-    return new Promise((resolve) => {
-      let baseURL = "";
-
-      let reader = new FileReader();
-
-      reader.readAsDataURL(file);
-
-      reader.onload = () => {
-        baseURL = reader.result;
-        resolve(baseURL);
-      };
-    });
-  };
-
-  const handleImage = (e) => {
-    let { file } = state;
-    file = e.target.files[0];
-
-    getBase64(file)
-      .then((result) => {
-        file["base64"] = result;
-        setFormData({
-          ...formData,
-          base64Url: result,
-          file,
-        });
-      })
-      .catch((err) => {
-        console.log("error", err);
-      });
-
-    setFormData({
-      ...state,
-      file: e.target.files[0],
-    });
-  };
-
-  // const handleSubmit = (e) => {
-  //   e.preventDefault();
-  //
-  //   let form_data = new FormData();
-  //   form_data.append("property_area", formData.property_area);
-  //   form_data.append("property_address", formData.property_address);
-  //   form_data.append("area_code", formData.area_code);
-  //   form_data.append("rent", formData.rent);
-  //   form_data.append("property_image", formData.base64Url);
-  //   let url = "/stocks/create/";
-  //
-  //   AxiosInstance.post(url, form_data, {
-  //     headers: {
-  //       "Content-Type": "multipart/form-data",
-  //     },
-  //   })
-  //     .then((res) => {
-  //       console.log("response", res);
-  //     })
-  //     .catch((err) => {
-  //       console.log("error", err);
-  //     });
-  // };
-
-  console.log("formData", formData);
-
-  return (
-    <>
-      {form ? (
-        <Form className={styles.container}>
-          <div className={styles.subheading}>
-            {form.loadForm ? (
-              <>
-                {subheading(form.subheadingText)}
-                {buttonClick(
-                  () => dispatch({ type: form.loadForm }),
-                  "here?",
-                  "link",
-                )}
-              </>
-            ) : (
-              <>{subheading(form.subheadingText)}</>
-            )}
-          </div>
-          {form.populateForm.map(
-            ({ index, label, type, placeholder, name }) => (
-              <Form.Group
-                className="mb-3"
-                controlId={`${name}${index}`}
-                key={index}
-              >
-                <Form.Label>{label}</Form.Label>
-                <Form.Control
-                  type={type}
-                  placeholder={placeholder}
-                  name={name}
-                  onChange={type === "file" ? handleImage : handleChange}
-                />
-              </Form.Group>
-            ),
-          )}
-          <div className={styles.subheading}>
-            {form.aboveBtnForm ? (
-              <>
-                {subheading(form.aboveBtnText)}
-                {buttonClick(
-                  () => dispatch({ type: form.aboveBtnForm }),
-                  "here?",
-                  "link",
-                )}
-              </>
-            ) : null}
-          </div>
-          <div className={styles.buttons}>
-            {buttonClick((e) => handleSubmit(e), form.buttonText, "success")}
-            {buttonClick(
-              () => dispatch({ type: "TOGGLE HIDE MODAL" }),
-              "Cancel",
-              "primary",
-            )}
-          </div>
-        </Form>
-      ) : null}
-    </>
-  );
+const FormOrganism = ({ auth }) => {
+  return <>{auth ? <AuthFormMolecule /> : <GeneralFormMolecule />}</>;
 };
 
 export default FormOrganism;
-
-// // 3rd party libraries
-// import Form from "react-bootstrap/Form";
-// import { useState } from "react";
-//
-// // components
-// import { buttonClick } from "../atom/elements";
-//
-// // styling
-// import styles from "../../styles/organism/Form.module.css";
-// import axiosDefaults, { axiosRes } from "../../API/axiosDefaults";
-//
-// const FormOrganism = ({ data }) => {
-//   const [state, setState] = useState({
-//     file: null,
-//     base64URL: "",
-//   });
-//   const [d, setD] = useState(null);
-//
-//   const handleChange = (e) => {
-//     e.preventDefault();
-//
-//     const { name, value } = e.target;
-//
-//     setState({
-//       ...state,
-//       [name]: value,
-//     });
-//   };
-//
-//   const handleSubmit = async (e) => {
-//     e.preventDefault();
-//
-//     try {
-//       const { data } = await axiosDefaults.post("/stocks/create", state);
-//       console.log("data", data);
-//       setD((prevD) => ({
-//         ...prevD,
-//         results: prevD.results.map((item) => {
-//           console.log("item", item);
-//         }),
-//       }));
-//     } catch (err) {
-//       console.log("error", err);
-//     }
-//   };
-//
-//   const getBase64 = (file) => {
-//     return new Promise((resolve) => {
-//       let baseURL = "";
-//
-//       let reader = new FileReader();
-//
-//       reader.readAsDataURL(file);
-//
-//       reader.onload = () => {
-//         baseURL = reader.result;
-//         resolve(baseURL);
-//       };
-//     });
-//   };
-//
-//   const handleImage = (e) => {
-//     let { file } = state;
-//     file = e.target.files[0];
-//
-//     getBase64(file)
-//       .then((result) => {
-//         file["base64"] = result;
-//         setState({
-//           ...state,
-//           base64Url: result,
-//           file,
-//         });
-//       })
-//       .catch((err) => {
-//         console.log("error", err);
-//       });
-//
-//     setState({
-//       ...state,
-//       file: e.target.files[0],
-//     });
-//   };
-//
-//   return (
-//     <Form className={styles.container}>
-//       {data.map((item, index) => (
-//         <Form.Group
-//           className="mb-3"
-//           key={index}
-//           controlId={`exampleForm.ControlInput${index}`}
-//         >
-//           <Form.Label>{item.label}</Form.Label>
-//           <Form.Control
-//             type={item.type}
-//             placeholder={item.placeholder}
-//             name={item.name}
-//             onChange={item.type === "file" ? handleImage : handleChange}
-//           />
-//         </Form.Group>
-//       ))}
-//       {buttonClick((e) => handleSubmit(e), "Save", "primary")}
-//     </Form>
-//   );
-// };
-//
-// export default FormOrganism;
