+++ src/utils/handlers/handleChange.jsx
@@ -1,0 +1,10 @@
+export const handleChange = (e, state, setState) => {
+  e.preventDefault();
+
+  const { name, value } = e.target;
+
+  setState({
+    ...state,
+    [name]: value,
+  });
+};
