+++ src/utils/handlers/index.jsx
@@ -1,0 +1,3 @@
+export * from "./handleChange";
+export * from "./handleImage";
+export * from "./handleSubmit";
